Removing Frame based Crypto provider
