+++ libs/encryption-sdk/src/main/java/org/opensearch/encryption/CryptoManagerFactory.java
@@ -15,8 +15,7 @@ import org.opensearch.common.crypto.CryptoProvider;
 import org.opensearch.common.crypto.MasterKeyProvider;
 import org.opensearch.common.unit.TimeValue;
 import org.opensearch.common.util.concurrent.AbstractRefCounted;
-import org.opensearch.encryption.frame.FrameCryptoProvider;
-import org.opensearch.encryption.frame.core.AwsCrypto;
+import org.opensearch.encryption.dummy.DummyCryptoProvider;
 import org.opensearch.encryption.keyprovider.CryptoMasterKey;
 
 import java.security.SecureRandom;
@@ -72,18 +71,8 @@ public class CryptoManagerFactory {
         MasterKeyProvider masterKeyProvider
     ) {
         switch (algorithm) {
-            case "ALG_AES_256_GCM_HKDF_SHA512_COMMIT_KEY":
-                return new FrameCryptoProvider(
-                    new AwsCrypto(materialsManager, CryptoAlgorithm.ALG_AES_256_GCM_HKDF_SHA512_COMMIT_KEY),
-                    masterKeyProvider.getEncryptionContext()
-                );
-            case "ALG_AES_256_GCM_HKDF_SHA512_COMMIT_KEY_ECDSA_P384":
-                return new FrameCryptoProvider(
-                    new AwsCrypto(materialsManager, CryptoAlgorithm.ALG_AES_256_GCM_HKDF_SHA512_COMMIT_KEY_ECDSA_P384),
-                    masterKeyProvider.getEncryptionContext()
-                );
             default:
-                throw new IllegalArgumentException("Unsupported algorithm: " + algorithm);
+                return new DummyCryptoProvider(materialsManager, masterKeyProvider);
         }
     }
